+++ app/src/public/java/com/mapbox/maps/testapp/examples/LocationComponentActivity.kt
@@ -1,5 +1,6 @@
 package com.mapbox.maps.testapp.examples
 
+import android.graphics.Color
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
@@ -7,9 +8,16 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.mapbox.maps.CameraOptions
 import com.mapbox.maps.ImageHolder
+import com.mapbox.maps.MapboxExperimental
 import com.mapbox.maps.Style
 import com.mapbox.maps.extension.style.expressions.dsl.generated.interpolate
+import com.mapbox.maps.extension.style.expressions.generated.Expression
+import com.mapbox.maps.extension.style.layers.addLayer
+import com.mapbox.maps.extension.style.layers.generated.fillExtrusionLayer
+import com.mapbox.maps.extension.style.layers.properties.generated.Anchor
 import com.mapbox.maps.extension.style.layers.properties.generated.ProjectionName
+import com.mapbox.maps.extension.style.light.generated.light
+import com.mapbox.maps.extension.style.light.generated.setLight
 import com.mapbox.maps.extension.style.projection.generated.getProjection
 import com.mapbox.maps.extension.style.projection.generated.projection
 import com.mapbox.maps.extension.style.projection.generated.setProjection
@@ -26,6 +34,7 @@ import com.mapbox.maps.testapp.databinding.ActivityLocationComponentBinding
 import com.mapbox.maps.testapp.utils.LocationPermissionHelper
 import java.lang.ref.WeakReference
 
+@OptIn(MapboxExperimental::class)
 class LocationComponentActivity : AppCompatActivity() {
 
   private var lastStyleUri = Style.DARK
@@ -46,6 +55,18 @@ class LocationComponentActivity : AppCompatActivity() {
     locationPermissionHelper.checkPermissions {
       binding.mapView.apply {
         getMapboxMap().loadStyleUri(Style.MAPBOX_STREETS) {
+          setupBuildings(it)
+          it.setLight(
+            light {
+              anchor(Anchor.MAP)
+              color(Color.YELLOW)
+              position(
+                radialCoordinate = 10.0,
+                azimuthalAngle = 40.0,
+                polarAngle = 50.0
+              )
+            }
+          )
           // Disable scroll gesture, since we are updating the camera position based on the indicator location.
           gestures.scrollEnabled = false
           gestures.addOnMapClickListener { point ->
@@ -73,6 +94,21 @@ class LocationComponentActivity : AppCompatActivity() {
         }
       }
     }
+  }
+
+  private fun setupBuildings(style: Style) {
+    val fillExtrusionLayer = fillExtrusionLayer("3d-buildings", "composite") {
+      sourceLayer("building")
+      filter(Expression.eq(Expression.get("extrude"), Expression.literal("true")))
+      minZoom(15.0)
+      fillExtrusionColor(Color.parseColor("#aaaaaa"))
+      fillExtrusionHeight(Expression.get("height"))
+      fillExtrusionBase(Expression.get("min_height"))
+      fillExtrusionOpacity(0.6)
+      fillExtrusionAmbientOcclusionIntensity(0.3)
+      fillExtrusionAmbientOcclusionRadius(3.0)
+    }
+    style.addLayer(fillExtrusionLayer)
   }
 
   override fun onCreateOptionsMenu(menu: Menu): Boolean {
@@ -174,9 +210,11 @@ class LocationComponentActivity : AppCompatActivity() {
         )
         is LocationPuck2D -> it.locationPuck = LocationPuck3D(
           modelUri = "asset://sportcar.glb",
-          modelScale = listOf(0.1f, 0.1f, 0.1f),
+          modelScale = listOf(10f, 10f, 10f),
           modelTranslation = listOf(0.1f, 0.1f, 0.1f),
           modelRotation = listOf(0.0f, 0.0f, 180.0f),
+          modelCastShadows = false,
+          modelReceiveShadows = false,
         )
       }
     }