+++ plugin-locationcomponent/src/public/java/com/mapbox/maps/plugin/locationcomponent/LayerSourceProvider.kt
@@ -28,7 +28,8 @@ internal class LayerSourceProvider {
       locationModelLayerOptions.modelScale.map { it.toDouble() },
       locationModelLayerOptions.modelRotation.map { it.toDouble() },
       locationModelLayerOptions.modelTranslation.map { it.toDouble() },
-      locationModelLayerOptions.modelOpacity.toDouble()
+      locationModelLayerOptions.modelOpacity.toDouble(),
+      locationModelLayerOptions.modelScaleMode
     )
 
   fun getLocationIndicatorLayer() = LocationIndicatorLayerWrapper(LOCATION_INDICATOR_LAYER)