+++ plugin-locationcomponent/src/public/java/com/mapbox/maps/plugin/locationcomponent/LocationPuckManager.kt
@@ -11,7 +11,6 @@ import com.mapbox.common.location.LocationError
 import com.mapbox.geojson.Point
 import com.mapbox.maps.MapboxLocationComponentException
 import com.mapbox.maps.Style
-import com.mapbox.maps.StylePropertyValueKind
 import com.mapbox.maps.logW
 import com.mapbox.maps.plugin.LocationPuck2D
 import com.mapbox.maps.plugin.LocationPuck3D
@@ -20,8 +19,6 @@ import com.mapbox.maps.plugin.locationcomponent.animators.PuckAnimatorManager
 import com.mapbox.maps.plugin.locationcomponent.generated.LocationComponentSettings
 import java.lang.ref.WeakReference
 import kotlin.math.abs
-import kotlin.math.cos
-import kotlin.math.pow
 
 internal class LocationPuckManager(
   var settings: LocationComponentSettings,
@@ -38,28 +35,8 @@ internal class LocationPuckManager(
   @VisibleForTesting(otherwise = PRIVATE)
   internal var lastLocation: Point? = null
 
-  @VisibleForTesting(otherwise = PRIVATE)
-  internal var lastMercatorScale = 1.0
-    set(value) {
-      if (abs(value - field) > MERCATOR_SCALE_THRESHOLD) {
-        field = value
-        (settings.locationPuck as? LocationPuck3D)?.let { locationPuck3D ->
-          locationLayerRenderer.styleScaling(get3DPuckScaleExpression(locationPuck3D, field))
-        }
-      }
-    }
-
   private val onLocationUpdated: ((Point) -> Unit) = {
     lastLocation = it
-    if (settings.locationPuck is LocationPuck3D) {
-      delegateProvider.getStyle { style ->
-        val latitude =
-          if (style.isGlobeProjection()) {
-            delegateProvider.mapCameraManagerDelegate.cameraState.center.latitude()
-          } else it.latitude()
-        lastMercatorScale = mercatorScale(latitude)
-      }
-    }
   }
 
   @VisibleForTesting(otherwise = PRIVATE)
@@ -266,102 +243,22 @@ internal class LocationPuckManager(
 
   /**
    * Function to set scaling for [LocationPuck].
-   * In order to keep 3D puck size constant across all zoom levels, we interpolate the model based on
-   * current zoom level. MIN_ZOOM, MAX_ZOOM are used as two anchor points to calculate
-   * the scale expression.
    */
   @VisibleForTesting(otherwise = PRIVATE)
   internal fun styleScaling(settings: LocationComponentSettings) {
     when (val puck = settings.locationPuck) {
       is LocationPuck2D -> {
-        val scaleExpression = puck.scaleExpression
-        if (scaleExpression != null) {
-          locationLayerRenderer.styleScaling(Value.fromJson(scaleExpression).take())
-        }
+        puck.scaleExpression
       }
       is LocationPuck3D -> {
-        locationLayerRenderer.styleScaling(get3DPuckScaleExpression(puck, lastMercatorScale))
+        puck.modelScaleExpression
       }
+    }?.let {
+      locationLayerRenderer.styleScaling(Value.fromJson(it).take())
     }
-  }
-
-  private fun get3DPuckScaleExpression(puck: LocationPuck3D, mercatorScale: Double): Value {
-    val modelScaleConstant = 2.0.pow(MAX_ZOOM - MIN_ZOOM)
-    val modelScaleExpression = puck.modelScaleExpression
-    return if (modelScaleExpression == null) {
-      Value(
-        arrayListOf(
-          Value("interpolate"),
-          Value(arrayListOf(Value("exponential"), Value(PUCK_3D_EXPONENTIAL_EXPRESSION_BASE))),
-          Value(arrayListOf(Value("zoom"))),
-          Value(MIN_ZOOM),
-          Value(
-            arrayListOf(
-              Value("literal"),
-              Value(
-                arrayListOf(
-                  Value(modelScaleConstant * puck.modelScale[0].toDouble() * mercatorScale),
-                  Value(modelScaleConstant * puck.modelScale[1].toDouble() * mercatorScale),
-                  Value(modelScaleConstant * puck.modelScale[2].toDouble() * mercatorScale)
-                )
-              )
-            )
-          ),
-          Value(MAX_ZOOM),
-          Value(
-            arrayListOf(
-              Value("literal"),
-              Value(
-                arrayListOf(
-                  Value(puck.modelScale[0].toDouble() * mercatorScale),
-                  Value(puck.modelScale[1].toDouble() * mercatorScale),
-                  Value(puck.modelScale[2].toDouble() * mercatorScale)
-                )
-              )
-            )
-          )
-        )
-      )
-    } else {
-      Value.fromJson(modelScaleExpression).take()
-    }
-  }
-
-  private fun mercatorScale(lat: Double): Double {
-    // In Mercator projection the scale factor is changed along the meridians as a function of latitude
-    // to keep the scale factor equal in all direction: k=sec(latitude), where sec(α) = 1 / cos(α).
-    // Here we are inverting the logic, as the 3d puck is using real-world size, and we are revising
-    // the appearance to look constant on a mercator projection map.
-    return cos(
-      // convert decimal latitude degrees to radians
-      lat.coerceIn(-LATITUDE_MAX, LATITUDE_MAX) * Math.PI / 180.0
-    )
-  }
-
-  private fun Style.isGlobeProjection(): Boolean {
-    val projectionProperty = getStyleProjectionProperty("name")
-    return projectionProperty.kind == StylePropertyValueKind.CONSTANT &&
-      (projectionProperty.value.contents as String).uppercase() == "GLOBE"
   }
 
   private companion object {
-    const val MIN_ZOOM = 0.50
-    const val MAX_ZOOM = 22.0
-
-    // To make the 3D puck's size constant across different zoom levels, the 3D puck's size (real world object size)
-    // should be exponential to the zoom level.
-    // The base of the exponential expression is decided by how the tile pyramid works: at zoom level n, we have 2^(n+1)
-    // tiles to cover the earth.
-    const val PUCK_3D_EXPONENTIAL_EXPRESSION_BASE = 0.5
-
-    // We display most of the world at the lowest zoom level as a single square image, excluding the
-    // polar regions by truncation at latitudes of φmax = ±85.05113°.
-    // refs: https://en.wikipedia.org/wiki/Mercator_projection#:~:text=Web%20Mercator,-Main%20article%3A%20Web&text=The%20major%20online%20street%20mapping,%CF%86max%20%3D%20%C2%B185.05113%C2%B0
-    const val LATITUDE_MAX = 85.051128779806604
-
-    // Threshold to update the mercator scale factor when the latitude changes, so that we don't update the
-    // scale expression too frequently and cause performance issues.
-    const val MERCATOR_SCALE_THRESHOLD = 0.01
     const val BEARING_UPDATE_THRESHOLD = 0.01
     const val TAG = "LocationPuckManager"
   }