+++ plugin-locationcomponent/src/testPublic/java/com/mapbox/maps/plugin/locationcomponent/LocationPuckManagerTest.kt
@@ -17,7 +17,6 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
-import kotlin.math.abs
 
 @RunWith(RobolectricTestRunner::class)
 class LocationPuckManagerTest {
@@ -295,97 +294,6 @@ class LocationPuckManagerTest {
     every { Value.fromJson(any()) } returns ExpectedFactory.createValue(Value("expression"))
     locationPuckManager.styleScaling(settings)
     verify { locationLayerRenderer.styleScaling(Value("expression")) }
-  }
-
-  @Test
-  fun testDefault3DStyleScaling() {
-    every { settings.locationPuck } returns LocationPuck3D(
-      modelUri = "uri",
-    )
-    every { Value.fromJson(any()) } returns ExpectedFactory.createValue(Value("expression"))
-    locationPuckManager.styleScaling(settings)
-    verify { locationLayerRenderer.styleScaling(capture(valueSlot)) }
-    val value = valueSlot.captured.toString()
-    assertTrue(value.startsWith("[interpolate, [exponential, 0.5], [zoom], 0.5, [literal, ["))
-    assertTrue(value.endsWith("]], 22.0, [literal, [1.0, 1.0, 1.0]]]"))
-    value
-      .replace("[interpolate, [exponential, 0.5], [zoom], 0.5, [literal, [", "")
-      .replace("]], 22.0, [literal, [1.0, 1.0, 1.0]]]", "")
-      .split(", ")
-      .map { it.toDouble() }
-      .forEach {
-        val absoluteDifference: Float = abs(it - MODEL_SCALE_CONSTANT).toFloat()
-        val maxUlp = Math.ulp(it).coerceAtLeast(Math.ulp(MODEL_SCALE_CONSTANT)).toFloat()
-        assert(absoluteDifference < 2 * maxUlp)
-      }
-  }
-
-  @Test
-  fun testDefault3DStyleScalingWithMercatorScale() {
-    // in mercator projection we take actual location update as camera center, so
-    // lat = 60.0, cos(60.0) = 0.5
-    testDefault3DScaling(
-      actualProjection = "mercator",
-      actualLocationUpdate = Point.fromLngLat(60.0, 60.0),
-      expectedLastMercatorScale = 0.5,
-      expectedLastMercatorScaleString = "0.5000000000000001"
-    )
-  }
-
-  @Test
-  fun testDefault3DStyleScalingWithGlobeScale() {
-    // in globe projection we take mapCameraDelegate.cameraState as camera center, so
-    // lat = 0.0, cos(0.0) = 1.0
-    testDefault3DScaling(
-      actualProjection = "globe",
-      actualLocationUpdate = Point.fromLngLat(60.0, 60.0),
-      expectedLastMercatorScale = 1.0,
-      expectedLastMercatorScaleString = "1.0"
-    )
-  }
-
-  private fun testDefault3DScaling(
-    actualProjection: String,
-    actualLocationUpdate: Point,
-    expectedLastMercatorScale: Double,
-    expectedLastMercatorScaleString: String
-  ) {
-    val onLocationUpdatedSlot = slot<((Point) -> Unit)>()
-    val getStyleSlot = slot<((Style) -> Unit)>()
-    val style = mockk<Style>()
-    every { style.getStyleProjectionProperty("name") } returns
-      StylePropertyValue(Value(actualProjection), StylePropertyValueKind.CONSTANT)
-    val valueSlots = mutableListOf<Value>()
-    every { settings.locationPuck } returns LocationPuck3D(
-      modelUri = "uri",
-    )
-    every { Value.fromJson(any()) } returns ExpectedFactory.createValue(Value("expression"))
-    locationPuckManager.initialize(mockk())
-    verify { animationManager.setUpdateListeners(capture(onLocationUpdatedSlot), any(), any()) }
-    onLocationUpdatedSlot.captured.invoke(actualLocationUpdate)
-    verify { delegateProvider.getStyle(capture(getStyleSlot)) }
-    getStyleSlot.captured.invoke(style)
-    assertEquals(expectedLastMercatorScale, locationPuckManager.lastMercatorScale, 1E-5)
-    verify { locationLayerRenderer.styleScaling(capture(valueSlots)) }
-    val value = valueSlots.last().toString()
-    assertTrue(value.startsWith("[interpolate, [exponential, 0.5], [zoom], 0.5, [literal, ["))
-    assertTrue(value.endsWith("]], 22.0, [literal, [$expectedLastMercatorScaleString, $expectedLastMercatorScaleString, $expectedLastMercatorScaleString]]]"))
-    value
-      .replace("[interpolate, [exponential, 0.5], [zoom], 0.5, [literal, [", "")
-      .replace(
-        "]], 22.0, [literal, [$expectedLastMercatorScaleString, $expectedLastMercatorScaleString, $expectedLastMercatorScaleString]]]",
-        ""
-      )
-      .split(", ")
-      .map { it.toDouble() }
-      .forEach {
-        val absoluteDifference: Float =
-          abs(it - MODEL_SCALE_CONSTANT * expectedLastMercatorScale).toFloat()
-        val maxUlp =
-          Math.ulp(it).coerceAtLeast(Math.ulp(MODEL_SCALE_CONSTANT * expectedLastMercatorScale))
-            .toFloat()
-        assert(absoluteDifference < 2 * maxUlp)
-      }
   }
 
   @Test